+++ bootstrap/src/test/java/com/onlydust/marketplace/indexer/bootstrap/it/RepoJobIndexingIT.java
@@ -17,7 +17,7 @@ public class RepoJobIndexingIT extends IntegrationTest {
     @Test
     public void should_add_repo_to_index() throws IOException {
         // Given
-        final Integer MARKETPLACE = 498695724;
+        final Long MARKETPLACE = 498695724L;
 
         // When
         final var response = indexRepo(MARKETPLACE);
@@ -25,10 +25,10 @@ public class RepoJobIndexingIT extends IntegrationTest {
         // Then
         response.expectStatus().isNoContent();
 
-        assertThat(repoIndexingJobTriggerRepository.list()).containsExactly(new RepoIndexingJobTrigger(0L, MARKETPLACE.longValue()));
+        assertThat(repoIndexingJobTriggerRepository.list()).containsExactly(new RepoIndexingJobTrigger(0L, MARKETPLACE));
     }
 
-    private WebTestClient.ResponseSpec indexRepo(Integer repoId) {
+    private WebTestClient.ResponseSpec indexRepo(Long repoId) {
         return put("/api/v1/indexes/repos/" + repoId);
     }
 }
